Clean code and math
1) Why is the change being made?
---------------------------------
Before the pong game cheated by simulating where the ball will end up.
This was prone to error and used more resources than necessary.
Additionally the logic was somewhat complicated and somewhat broke
firefox support.

2) What has changed?
---------------------------------
Now a group theoretic approach is taken to precisely calculate the
end-posiiton of the ball. The intuition is that the direction of the
ball can be encoded in the position--this makes the velocity stay
constant so it is easy to predict where the ball will end up (despite
the ball bouncing up/down constantly in reality).

3) Other concerns and/or dependent changes?
---------------------------------
Most importantly, the index.html file is a total mess. Some js files
should be added.

The ball start velocity is randomly calculated at the start of each round, this works, but most pong games will change the speed+velocity depending on where in the racket the ball hits. This should likely be added in the future.